Se cambia tabla SplitterSpacerChart2

+++ DProcess/DataAccess/ServiceObjects/Tooling/SPlitterSpacerChart2.cs
@@ -19,7 +19,7 @@ namespace DataAccess.ServiceObjects.Tooling
         public Nullable<double> SplitWidth { get; set; }
         public Nullable<double> GrindWidth { get; set; }
         public double CutterSpacer1 { get; set; }
-        public Nullable<double> CantidadSpacer1 { get; set; }
+        public Nullable<int> CantidadSpacer1 { get; set; }
         public Nullable<double> CutterSpacer2 { get; set; }
         public int CantidadSpacer2 { get; set; }
         public int ident { get; set; }
